implemented resourcesview
