+++ app/src/main/java/org/techconnect/views/ResourcesView.java
@@ -1,11 +1,17 @@
 package org.techconnect.views;
 
 import android.content.Context;
+import android.content.Intent;
 import android.util.AttributeSet;
+import android.view.View;
+import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import org.centum.techconnect.R;
+import org.techconnect.activities.PDFActivity;
+import org.techconnect.misc.Utils;
+import org.techconnect.resources.ResourceHandler;
 
 import java.util.List;
 
@@ -41,11 +47,38 @@ public class ResourcesView extends LinearLayout {
     }
 
     private void updateViews() {
+        removeAllViews();
+        addView(resourcesHeaderTextView);
         if (resources.size() == 0) {
             resourcesHeaderTextView.setText(R.string.resources_none_msg);
         } else {
             resourcesHeaderTextView.setText(R.string.resources_general_msg);
         }
+
+        for (final String att : resources) {
+            String name = Utils.formatAttachmentName(att);
+            Button button = new Button(getContext());
+            button.setTransformationMethod(null);
+            button.setText(name);
+            button.setOnClickListener(new OnClickListener() {
+                @Override
+                public void onClick(View view) {
+                    openAttachment(att);
+                }
+            });
+            addView(button);
+        }
+    }
+
+    private void openAttachment(String att) {
+        Intent intent = new Intent(getContext(), PDFActivity.class);
+        intent.putExtra(PDFActivity.EXTRA_IS_FILE, true);
+        if (ResourceHandler.get().hasStringResource(att)) {
+            intent.putExtra(PDFActivity.EXTRA_FILE, getContext().getFileStreamPath(ResourceHandler.get().getStringResource(att)).getAbsolutePath());
+        } else {
+            intent.putExtra(PDFActivity.EXTRA_FILE, "");
+        }
+        getContext().startActivity(intent);
     }
 
     @Override
